tests/wpt/web-platform-tests/css/css-ui/caret-color-021.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Basic User Interface Test: Default caret-color test animation</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="http://www.w3.org/TR/css3-ui/#caret-color">
<link rel="help" href="https://www.w3.org/TR/css3-animations/#animation">
<link rel="help" href="https://www.w3.org/TR/css3-color/#color0">
<meta name="assert" content="Test checks that the default value for caret-color property, which is 'auto', is not interpolable.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes caret-color-to-lime {
    to { caret-color: lime; }
  }

  #textarea {
    color: magenta;
    animation: caret-color-to-lime 2s -1s paused;
  }
</style>
<body>
  <textarea id="textarea"></textarea>
  <div id=log></div>

  <script>
    test(
      function(){
        var textarea = document.getElementById("textarea");
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(0, 255, 0)');
      }, "Default caret-color is not interpolable");
</script>
</body>
