LayoutTests/imported/w3c/web-platform-tests/css/css-ui/caret-color-021-expected.txt



FAIL Default caret-color is not interpolable assert_equals: expected "rgb(0, 255, 0)" but got "rgba(0, 255, 0, 0.803922)"
